+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Add document upload API
 * Add AuthorizationAdjustment class and `authorizationAdjustments` to Transaction
 * Coinbase is no longer a supported payment method. `PAYMENT_METHOD_NO_LONGER_SUPPORTED` will be returned for Coinbase operations.
-* Add facilitator details to Transaction if appropriate
+* Add facilitated transaction details to Transaction if present
 
 ## 2.72.1